+++ src/ClinicApp.Domain/ClinicApp.Domain.csproj
@@ -11,8 +11,4 @@
     <PackageReference Include="MediatR" Version="13.0.0" />
   </ItemGroup>
 
-  <ItemGroup>
-    <Folder Include="Common\Interfaces\" />
-  </ItemGroup>
-
 </Project>
