+++ src/ClinicApp.Domain/Common/Interfaces/IClock.cs
@@ -1,0 +1,6 @@
+﻿namespace ClinicApp.Domain.Common.Interfaces;
+
+public interface IClock
+{
+    public DateTime UtcNow => DateTime.UtcNow;
+}
