+++ src/ClinicApp.Domain/Common/SystemClock.cs
@@ -1,0 +1,8 @@
+﻿using ClinicApp.Domain.Common.Interfaces;
+
+namespace ClinicApp.Domain.Common;
+
+public class SystemClock : IClock
+{
+    public DateTime UtcNow => DateTime.UtcNow;
+}
