--- src/ClinicApp.Domain/Doctor/WorkingDays.cs
@@ -1,16 +1,0 @@
-﻿namespace ClinicApp.Domain.Doctor;
-
-[Flags]
-public enum WorkingDays
-{
-    None = 0,
-    Sunday = 1 << 0,
-    Monday = 1 << 1,
-    Tuesday = 1 << 2,
-    Wednesday = 1 << 3,
-    Thursday = 1 << 4,
-    Friday = 1 << 5,
-    Saturday = 1 << 6 
-}
-
-
