+++ src/ClinicApp.Domain/DoctorAgg/WorkingDays.cs
@@ -1,0 +1,16 @@
+﻿namespace ClinicApp.Domain.DoctorAgg;
+
+[Flags]
+public enum WorkingDays
+{
+    None = 0,
+    Sunday = 1 << 0,
+    Monday = 1 << 1,
+    Tuesday = 1 << 2,
+    Wednesday = 1 << 3,
+    Thursday = 1 << 4,
+    Friday = 1 << 5,
+    Saturday = 1 << 6 
+}
+
+
