+++ src/ClinicApp.Domain/Repositories/IDoctorRepository.cs
@@ -1,10 +1,10 @@
-using ClinicApp.Domain.Doctor;
+using ClinicApp.Domain.DoctorAgg;
 
 namespace ClinicApp.Domain.Repositories
 {
     public interface IDoctorRepository
     {
-        Task<Doctor.Doctor?> GetByIdAsync(Guid id);
-        void Add(Doctor.Doctor doctor);
+        Task<Doctor?> GetByIdAsync(Guid id);
+        void Add(Doctor doctor);
     }
 }
