+++ src/ClinicApp.Infrastructure/ClinicApp.Infrastructure.csproj
@@ -9,6 +9,12 @@
   <ItemGroup>
     <PackageReference Include="ErrorOr" Version="2.0.1" />
     <PackageReference Include="MediatR" Version="13.0.0" />
+    <PackageReference Include="Microsoft.AspNetCore.Http.Abstractions" Version="2.3.0" />
+    <PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL" Version="8.0.0" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\ClinicApp.Domain\ClinicApp.Domain.csproj" />
   </ItemGroup>
 
 </Project>
