+++ src/ClinicApp.Infrastructure/Configurations/RoomEntityConfiguration.cs
@@ -1,5 +1,5 @@
 ﻿using ClinicApp.Domain.Common.Entities;
-using ClinicApp.Domain.Session;
+using ClinicApp.Domain.SessionAgg;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
