+++ tests/ClinicApp.Domain.Tests.UnitTest/FakerClock.cs
@@ -1,0 +1,9 @@
+﻿using ClinicApp.Domain.Common.Interfaces;
+
+namespace ClinicApp.Domain.Tests.UnitTest;
+
+internal class FakerClock : IClock
+{
+    public DateTime UtcNow { get; set; }
+
+}
