+++ tests/ClinicApp.Domain.Tests.UnitTest/TemporaryTests/DoctorSessionSchedulerTests.cs
@@ -1,8 +1,7 @@
 using Moq;
 using Xunit;
 using FluentAssertions;
-using ClinicApp.Domain.Doctor;
-using ClinicApp.Domain.Session;
+using ClinicApp.Domain.DoctorAgg;
 using ClinicApp.Domain.Repositories;
 using ClinicApp.Domain.Common.ValueObjects;
 using ClinicApp.Domain.Services.Sessions;
@@ -10,6 +9,7 @@ using ErrorOr;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using System;
+using ClinicApp.Domain.SessionAgg;
 
 namespace ClinicApp.Domain.Tests.UnitTest.TemporaryTests
 {
@@ -28,11 +28,12 @@ namespace ClinicApp.Domain.Tests.UnitTest.TemporaryTests
         public async Task CreateSession_Should_ReturnError_WhenSessionIsOutsideOfWorkingDays()
         {
             // Arrange
-            var doctor = Factories.DoctorFactory();
+            var doctor = Factories.DoctorFactory;
+            var fakerClock = new FakerClock { UtcNow =  new DateTime(2025,7,21)};
             var sessionTime = TimeRange.Create(new DateTime(2025, 7, 22), new DateTime(2025, 7, 22, 10, 30, 0)).Value; // Tuesday
-            var session = Session.Session.Create(Guid.NewGuid(), sessionTime, new SessionDescription("Test"), Guid.NewGuid(), Guid.NewGuid(), doctor.Id).Value;
+            var session = Session.Create(Guid.NewGuid(), sessionTime, new SessionDescription("Test"), Guid.NewGuid(), Guid.NewGuid(), doctor.Id,fakerClock).Value;
 
-            _sessionRepoMock.Setup(repo => repo.GetAllSessionsForDoctor(doctor)).ReturnsAsync(new List<Session.Session>());
+            _sessionRepoMock.Setup(repo => repo.GetAllSessionsForDoctor(doctor)).ReturnsAsync(new List<Session>());
 
             // Act
             var result = await _scheduler.CreateSession(session, doctor);
@@ -46,17 +47,18 @@ namespace ClinicApp.Domain.Tests.UnitTest.TemporaryTests
         public async Task CreateSession_Should_ReturnError_WhenSessionsOverlap()
         {
             // Arrange
-            var doctor = Factories.DoctorFactory();
+            var doctor = Factories.DoctorFactory;
 
+            var fakerClock = new FakerClock { UtcNow = new DateTime(2025, 7, 21) };
 
             var existingSessionTime = TimeRange.Create(new DateTime(2025, 7, 21, 10, 0, 0), new DateTime(2025, 7, 21, 11, 0, 0)).Value;
-            var existingSession = Session.Session.Create(Guid.NewGuid(), existingSessionTime, new SessionDescription("Existing"), Guid.NewGuid(), Guid.NewGuid(), doctor.Id).Value;
+            var existingSession = Session.Create(Guid.NewGuid(), existingSessionTime, new SessionDescription("Existing"), Guid.NewGuid(), Guid.NewGuid(), doctor.Id,fakerClock).Value;
 
             var newSessionTime = TimeRange.Create(new DateTime(2025, 7, 21, 10, 30, 0), new DateTime(2025, 7, 21, 11, 30, 0)).Value;
-            var newSession = Session.Session.Create(Guid.NewGuid(), newSessionTime, new SessionDescription("New"), Guid.NewGuid(), Guid.NewGuid(), doctor.Id).Value;
+            var newSession = Session.Create(Guid.NewGuid(), newSessionTime, new SessionDescription("New"), Guid.NewGuid(), Guid.NewGuid(), doctor.Id, fakerClock).Value;
 
-            _sessionRepoMock.Setup(repo => repo.GetAllSessionsForDoctor(doctor)).ReturnsAsync(new List<Session.Session>());
-            _sessionRepoMock.Setup(repo => repo.GetFutureSessionsDoctor(doctor)).ReturnsAsync(new List<Session.Session> { existingSession });
+            _sessionRepoMock.Setup(repo => repo.GetAllSessionsForDoctor(doctor)).ReturnsAsync(new List<Session>());
+            _sessionRepoMock.Setup(repo => repo.GetFutureSessionsDoctor(doctor)).ReturnsAsync(new List<Session> { existingSession });
 
             // Act
             var result = await _scheduler.CreateSession(newSession, doctor);
@@ -70,13 +72,15 @@ namespace ClinicApp.Domain.Tests.UnitTest.TemporaryTests
         public async Task CreateSession_Should_Succeed_WhenSlotIsAvailable()
         {
             // Arrange
-            var doctor = Factories.DoctorFactory();
+            var doctor = Factories.DoctorFactory;
 
             var sessionTime = TimeRange.Create(new DateTime(2025, 7, 21, 14, 0, 0), new DateTime(2025, 7, 21, 15, 0, 0)).Value;
-            var session = Session.Session.Create(Guid.NewGuid(), sessionTime, new SessionDescription("Test"), Guid.NewGuid(), Guid.NewGuid(), doctor.Id).Value;
+            var fakerClock = new FakerClock { UtcNow = new DateTime(2025, 7, 21) };
 
-            _sessionRepoMock.Setup(repo => repo.GetAllSessionsForDoctor(doctor)).ReturnsAsync(new List<Session.Session>());
-            _sessionRepoMock.Setup(repo => repo.GetFutureSessionsDoctor(doctor)).ReturnsAsync(new List<Session.Session>());
+            var session = Session.Create(Guid.NewGuid(), sessionTime, new SessionDescription("Test"), Guid.NewGuid(), Guid.NewGuid(), doctor.Id, fakerClock).Value;
+
+            _sessionRepoMock.Setup(repo => repo.GetAllSessionsForDoctor(doctor)).ReturnsAsync(new List<Session>());
+            _sessionRepoMock.Setup(repo => repo.GetFutureSessionsDoctor(doctor)).ReturnsAsync(new List<Session>());
 
             // Act
             var result = await _scheduler.CreateSession(session, doctor);
