+++ tests/ClinicApp.Domain.Tests.UnitTest/TemporaryTests/SessionTests.cs
@@ -1,22 +1,30 @@
 
-using ClinicApp.Domain.Session;
 using ClinicApp.Domain.Common.ValueObjects;
 using Xunit;
 using FluentAssertions;
+using ClinicApp.Domain.SessionAgg;
+using ClinicApp.Domain.Common.Interfaces;
+using ErrorOr;
 
 namespace ClinicApp.Domain.Tests.UnitTest.TemporaryTests
 {
     public class SessionTests
     {
+        private IClock clock;
+        public SessionTests()
+        {
+            clock = new FakerClock() { UtcNow = new DateTime(2025, 11, 2) };
+        }
+        private ErrorOr<Session> GetSession(TimeRange sessionTime) => Session.Create(Guid.NewGuid(), sessionTime, new SessionDescription("Test"), Guid.NewGuid(), Guid.NewGuid(), Guid.NewGuid(), clock);
         [Fact]
         public void Create_Should_ReturnError_When_SessionIsInThePast()
         {
             // Arrange
-            var sessionTime = TimeRange.Create(DateTime.UtcNow.AddDays(-1), DateTime.UtcNow.AddDays(-1).AddHours(1)).Value;
+            var sessionTime = TimeRange.Create(clock.UtcNow.AddDays(-1), clock.UtcNow.AddDays(-1).AddHours(1)).Value;
 
             // Act
-            var result = Session.Session.Create(Guid.NewGuid(), sessionTime, new SessionDescription("Test"), Guid.NewGuid(), Guid.NewGuid(), Guid.NewGuid());
-
+            var result = GetSession(sessionTime);
+            
             // Assert
             result.IsError.Should().BeTrue();
             result.FirstError.Code.Should().Be(SessionErrors.SessionTimeInThePast.code);
@@ -26,8 +34,8 @@ namespace ClinicApp.Domain.Tests.UnitTest.TemporaryTests
         public void Session_Should_TransitionThroughLifecycleCorrectly()
         {
             // Arrange
-            var sessionTime = TimeRange.Create(DateTime.UtcNow.AddDays(1), DateTime.UtcNow.AddDays(1).AddHours(1)).Value;
-            var session = Session.Session.Create(Guid.NewGuid(), sessionTime, new SessionDescription("Test"), Guid.NewGuid(), Guid.NewGuid(), Guid.NewGuid()).Value;
+            var sessionTime = TimeRange.Create(clock.UtcNow.AddDays(1), clock.UtcNow.AddDays(1).AddHours(1)).Value;
+            var session = GetSession(sessionTime).Value;
 
             // Act & Assert
             session.SessionStatus.Should().Be(SessionStatus.Pending);
@@ -46,14 +54,14 @@ namespace ClinicApp.Domain.Tests.UnitTest.TemporaryTests
         public void UpdateDate_Should_ReturnError_When_SessionIsFinished()
         {
             // Arrange
-            var sessionTime = TimeRange.Create(DateTime.UtcNow.AddDays(1), DateTime.UtcNow.AddDays(1).AddHours(1)).Value;
-            var session = Session.Session.Create(Guid.NewGuid(), sessionTime, new SessionDescription("Test"), Guid.NewGuid(), Guid.NewGuid(), Guid.NewGuid()).Value;
+            var sessionTime = TimeRange.Create(clock.UtcNow.AddDays(1), clock.UtcNow.AddDays(1).AddHours(1)).Value;
+            var session = GetSession(sessionTime).Value;
             session.SetSession();
             session.StartSession();
             session.FinishSession();
 
             // Act
-            var newTime = TimeRange.Create(DateTime.UtcNow.AddDays(2), DateTime.UtcNow.AddDays(2).AddHours(1)).Value;
+            var newTime = TimeRange.Create(clock.UtcNow.AddDays(2), clock.UtcNow.AddDays(2).AddHours(1)).Value;
             var result = session.UpdateDate(newTime);
 
             // Assert
